add DepthGradedMultilayer type, density in constructors & Elements
Elements now have a density field as well, which can be handed to the
`init` call (in the future we might look them up as well).

Compounds can also be given a density in the `compound` macro by
providing a `ρ` or `density` argument.

The `DepthGradedMultilayer` type takes care of constructing the
correct refractive indices & layers for such a coating.