add `staticBuild` option and default to runtime elements
Two important changes:

1. Previous to this commit our default were compile time elements with
an option to convert those to runtime values. This had some negative
side effects, making performance much worse when dealing with many
elements. Now, we do all the hard work at runtime and simply have a
thin compile time safe wrapper around the elements with a static size
argument.

2. This adds the `-d:staticBuild` option, which slurps all the data
files in the `resources` directory so that we can have an independent
build that includes all data files. This is very handy when compiling
with musl for a fully static executable.